Broke CartesianProductFunction.pack() into packBinary(), packUnary(), and packLexical().
Fixed a few unit tests.

+++ java/src/edu/ohsu/cslu/parser/CsrSpmvPerMidpointParser.java
@@ -116,7 +116,7 @@ public class CsrSpmvPerMidpointParser extends CsrSpmvParser {
 
             for (int j = rightCell.offset(); j <= rightCell.maxRightChildIndex(); j++) {
 
-                final int childPair = cpf.pack(leftChild, (short) nonTerminalIndices[j]);
+                final int childPair = cpf.pack(leftChild, nonTerminalIndices[j]);
                 if (!cpf.isValid(childPair)) {
                     continue;
                 }
